pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nashvec"
version = "0.1.0"
description = "Optimization of Latent-Space Compression using Game-Theoretic Techniques for Transformer-Based Vector Search"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Kushagra Agrawal, Nisharg Nargund, Oishani Banerjee", email = "kushagraagrawal@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "tensorflow>=2.10.0",
    "numpy>=1.25.0",
    "scikit-learn>=1.0.0",
    "pandas>=2.0.0",
    "sentence-transformers>=2.0.0",
    "faiss-cpu>=1.12.0",
    "hnswlib>=0.8.0",
    "datasets>=2.0.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
nashvec-train = "nashvec.cli:train"
nashvec-query = "nashvec.cli:query"
nashvec-benchmark = "nashvec.cli:benchmark"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

